edit modal, getListManufacturer

+++ Project/Admin/AdminPage.js
@@ -1,14 +1,14 @@
 import { ProductMethod } from "./ProductMethod.js";
 
-let productTableBody = document.getElementById("productTableBody");
-
 let listProduct = ProductMethod.getListProductFromLocalStorage();
-displayProductOnAdminPage(listProduct);
+
 
 const btnSave = document.getElementById("btnSave");
 btnSave.addEventListener("click", function () {
     let listProduct = ProductMethod.addNewProduct();
     displayProductOnAdminPage(listProduct);
+    displayManufacturer();
+    displayCategory();
 });
 
 const btnSearch = document.getElementById("btnSearch");
@@ -22,12 +22,92 @@ btnSearch.addEventListener("click", function () {
 let btnEditSave = document.getElementById("btnEditSave");
 btnEditSave.addEventListener("click", function () {
     let listProduct = ProductMethod.saveEditProduct();
-
     displayProductOnAdminPage(listProduct);
+    displayManufacturer();
+    displayCategory();
 });
+
+let btnAddnew = document.getElementById("btnAddNew");
+btnAddnew.addEventListener("click", function () {
+    // let list = Array.from(document.getElementsByTagName("input"));
+    // console.log(list);
+    // for(var i = 0; i < list.length; i ++){
+    //     list[i].value = "";
+    // }
+    Array.from(document.getElementsByTagName("input")).forEach(element => {
+        element.value = "";
+    }); // clean form
+
+    displayCategoryForm();
+    displayManufacturerForm();
+})
+
+function displayManufacturer() {
+    let listManufacturer = document.getElementById("listManufacturer");
+    listManufacturer.innerHTML = "";
+    let manufacturerList = ProductMethod.getManufacturerList();
+    manufacturerList.forEach(element => {
+        listManufacturer.innerHTML +=
+            `<p class="list-group-item">
+                <button type="button" id="`+ element + `" class="btnClass">` + element + `</button>
+            </p>`;
+    });
+    manufacturerList.forEach(element => {
+        let btnManufacturer = document.getElementById(element);
+        btnManufacturer.addEventListener("click", function () {
+            let currentList = ProductMethod.searchProductByManufacturer(element)
+            displayProductOnAdminPage(currentList);
+        });
+    });
+}
+displayManufacturer();
+
+function displayCategory() {
+    let categoryList = ProductMethod.getCategoryList();
+    let listCategory = document.getElementById("listCategory");
+    listCategory.innerHTML = "";
+    categoryList.forEach(element => {
+        listCategory.innerHTML +=
+            '<option>' + element + '</option>'
+    });
+}
+displayCategory();
+
+
+function displayManufacturerForm() {
+    let manufacturerList = ProductMethod.getManufacturerList();
+    let datalist = document.getElementById("listManufacturerForm");
+    datalist.innerHTML = "";
+    console.log(manufacturerList);
+    manufacturerList.forEach(element => {
+        var option = document.createElement("option");
+        option.setAttribute("value", element);
+        datalist.appendChild(option);
+    });
+}
+
+function displayCategoryForm() {
+    let categoryList = ProductMethod.getCategoryList();
+    const datalist = document.getElementById("listCategoryForm");
+    // for (var i = 0 ; i < categoryList.length ; i++){
+    //     var option = document.createElement("option");
+    //     option.setAttribute("value" , categoryList[i]);
+    //     console.log(categoryList[i]);
+    //     datalist.appendChild(option);
+    // }
+
+    categoryList.forEach(element => {
+        var option = document.createElement("option");
+        option.setAttribute("value", element);
+        datalist.appendChild(option);
+    });
+}
+displayCategoryForm();
+
 
 
 function displayProductOnAdminPage(listProduct) {
+    let productTableBody = document.getElementById("productTableBody");
     productTableBody.innerHTML = "";
     var count = 0;
     listProduct.forEach(element => {
@@ -35,6 +115,7 @@ function displayProductOnAdminPage(listProduct) {
         productTableBody.innerHTML +=
             '<tr id="' + element.getID + '">' +
             '<td>' + count + '</td>' +
+            '<td>' + element.getID + '</td>' +
             '<td>' + element.getName + '</td>' +
             '<td>' + element.getPrice + '</td>' +
             '<td>' + element.getInfo + '</td>' +
@@ -56,9 +137,12 @@ function displayProductOnAdminPage(listProduct) {
         btnDelete.setAttribute("class", "btn btn-danger");
         btnDelete.innerText = "Delete";
         btnDelete.addEventListener("click", function () {
-            let listProduct = ProductMethod.deleteProduct(element.getID);
-            displayProductOnAdminPage(listProduct);
-        })
+            if (confirm("Are you sure to delete product with ID " + element.getID)) {
+                let listProduct = ProductMethod.deleteProduct(element.getID);
+                displayProductOnAdminPage(listProduct);
+                displayManufacturer();
+            }
+        });
 
         let btnEdit = document.createElement("button");
         btnEdit.setAttribute("type", "button");
@@ -76,3 +160,20 @@ function displayProductOnAdminPage(listProduct) {
         tr.appendChild(td_delete);
     });
 }
+
+displayProductOnAdminPage(listProduct);
+
+let btnChange = document.getElementById("btnTest");
+btnChange.addEventListener("click", () => {
+    let test = document.getElementById("test");
+    console.dir(test);
+});
+
+let test = [1, 2, 3];
+let result = 0;
+for(let i = 0; i < test.length; i ++){
+    if(i % 2 == 0){
+        result += test[i];
+    }
+}
+document.getElementById("test").innerHTML = i;
